+++ src/main/java/kr/sofaware/slas/auth/SecurityConfig.java
@@ -21,7 +21,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
         http
             .authorizeRequests()
-                .antMatchers("/favicon.ico", "/login", "/signup").permitAll() // 누구나 접근 가능
+                .antMatchers("/login", "/signup").permitAll() // 누구나 접근 가능
                 .antMatchers("/p/**").hasRole("PROFESSOR")
                 .antMatchers("/s/**").hasRole("STUDENT")
                 .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll() // 정적 리소스 항상 공개
