+++ src/main/java/kr/sofaware/slas/lecture/service/LectureService.java
@@ -1,0 +1,11 @@
+package kr.sofaware.slas.lecture.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class LectureService {
+
+
+}
