--- src/main/java/kr/sofaware/slas/service/LectureService.java
@@ -1,17 +1,0 @@
-package kr.sofaware.slas.service;
-
-import kr.sofaware.slas.entity.Lecture;
-import kr.sofaware.slas.repo.LectureRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@RequiredArgsConstructor
-@Service
-public class LectureService {
-
-
-}
