+++ src/main/java/kr/sofaware/slas/timetable/controller/TimetableProfessorController.java
@@ -1,8 +1,6 @@
-package kr.sofaware.slas.controller;
+package kr.sofaware.slas.timetable.controller;
 
-import kr.sofaware.slas.service.LectureService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Controller;
@@ -17,11 +15,8 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @Controller
-@RequestMapping("/s")
-public class TimetableStudentController {
-
-    @Autowired
-    private final LectureService lectureService;
+@RequestMapping("/p")
+public class TimetableProfessorController {
 
     @GetMapping("timetable")
     public String getStudentTimetable(Authentication authentication, Principal principal, Model model){
@@ -30,10 +25,8 @@ public class TimetableStudentController {
         String userId = principal.getName();
 
         model.addAttribute("id", userId);
-        model.addAttribute("auth",auth);
+        model.addAttribute("auth", auth);
 
-        //lecture service
-        //model.addAttribute("lectures", lectureService.findAll());
 
         return "timetable";
     }
