+++ src/main/java/kr/sofaware/slas/timetable/controller/TimetableStudentController.java
@@ -1,6 +1,8 @@
-package kr.sofaware.slas.controller;
+package kr.sofaware.slas.timetable.controller;
 
+import kr.sofaware.slas.lecture.service.LectureService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Controller;
@@ -9,14 +11,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.security.Principal;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 
 @RequiredArgsConstructor
 @Controller
-@RequestMapping("/p")
-public class TimetableProfessorController {
+@RequestMapping("/s")
+public class TimetableStudentController {
+
+    @Autowired
+    private final LectureService lectureService;
 
     @GetMapping("timetable")
     public String getStudentTimetable(Authentication authentication, Principal principal, Model model){
@@ -25,8 +28,10 @@ public class TimetableProfessorController {
         String userId = principal.getName();
 
         model.addAttribute("id", userId);
-        model.addAttribute("auth", auth);
+        model.addAttribute("auth",auth);
 
+        //lecture service
+        //model.addAttribute("lectures", lectureService.findAll());
 
         return "timetable";
     }
